chain groups onboarding
