+++ .env
@@ -2,9 +2,11 @@
 VITE_REACT_APP_PAYMENT_CANISTER_ID="ulvla-h7777-77774-qaacq"
 
 # DFX CANISTER ENVIRONMENT VARIABLES
-DFX_VERSION='0.27.0'
+DFX_VERSION='0.28.0'
 DFX_NETWORK='local'
-CANISTER_ID_ICP_LEDGER_CANISTER='uxrrr-q7777-77774-qaaaq-cai'
-CANISTER_ID='uxrrr-q7777-77774-qaaaq-cai'
-CANISTER_CANDID_PATH='/mnt/c/rogetz-folder/RotateChain_new/.dfx/local/canisters/icp_ledger_canister/icp_ledger_canister.did'
+CANISTER_ID_INTERNET_IDENTITY='rdmx6-jaaaa-aaaaa-aaadq-cai'
+CANISTER_ID_ICP_LEDGER_CANISTER='u6s2n-gx777-77774-qaaba-cai'
+CANISTER_ID_CHAIN_MANAGEMENT='uzt4z-lp777-77774-qaabq-cai'
+CANISTER_ID='uzt4z-lp777-77774-qaabq-cai'
+CANISTER_CANDID_PATH='/mnt/d/rogetz-folder/rotatechain_new/.dfx/local/canisters/chain_management/chain_management.did'
 # END DFX CANISTER ENVIRONMENT VARIABLES
