+++ dfx.json
@@ -22,7 +22,12 @@
       "type": "custom",
       "candid": "https://github.com/dfinity/ic/releases/download/ledger-suite-icp-2025-07-04/ledger.did",
       "wasm": "https://github.com/dfinity/ic/releases/download/ledger-suite-icp-2025-07-04/ledger-canister_notify-method.wasm.gz",
-      "init_arg" : "(variant { Init = record { minting_account = \"c571e55d926f0a9dda0b3820338649a980fbdee36915476add1a7944f53f6b1d\"; initial_values = vec { record {\"b2f6fd9f822ff1ca2ae79b04f24f8d184495d906f0b1e3e4990017830f2fd6c9\"; record { e8s = 10_000_000_000 : nat64; }; }; }; send_whitelist = vec {}; transfer_fee = opt record { e8s = 10_000 : nat64; }; token_symbol = opt \"LICP\"; token_name = opt \"Local ICP\"; } })"
+      "remote": {
+        "id": {
+          "ic": "ryjl3-tyaaa-aaaaa-aaaba-cai"
+        }
+      },
+      "init_arg" : "(variant { Init = record { minting_account = \"9e4e9642c7c6260f1fd93396fd3b5ee7585d0e16a03621228139d823a39540b2\"; initial_values = vec { record {\"3a701f682617e222368b81b998b994110cb735f25a95bff4117fea7dfbe0ecb6\"; record { e8s = 10_000_000_000 : nat64; }; }; }; send_whitelist = vec {}; transfer_fee = opt record { e8s = 10_000 : nat64; }; burn_fee = opt record { e8s = 0 : nat64; }; token_symbol = opt \"LICP\"; token_name = opt \"Local ICP\"; } })"
     },
     "icp_backend":{
       "main": "src/RotateChain_backend/icp_integration.mo",
