+++ src/RotateChain_frontend/src/components/sassySplash.tsx
