+++ src/RotateChain_frontend/src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './app_2'; // Make sure to import your App component
+import App from './app'; // Make sure to import your App component
 import "./globals.css"
 
 import {store} from './state/store'
@@ -20,16 +20,3 @@ ReactDOM.createRoot(document.getElementById('root')).render(
   </React.StrictMode>,
 );
 
-
-/*const rootElement = document.getElementById('root');
-
-if (rootElement) {
-  const root = ReactDOM.createRoot(rootElement);
-  root.render(
-    <React.StrictMode>
-      <App />
-    </React.StrictMode>
-  );
-} else {
-  console.error('Failed to find the root element');
-}*/
